requirements.txt
Flask
google-search-results
python-dotenv